Stop populating public package bucket (replaced by exported-api)
