+++ app/config/dartlang-pub.yaml
@@ -4,7 +4,6 @@
 # Storage is retrieved from the Datastore.
 ---
 canonicalPackagesBucketName: dartlang-pub-canonical-packages
-publicPackagesBucketName: dartlang-pub-public-packages
 incomingPackagesBucketName: dartlang-pub-incoming-packages
 projectId: dartlang-pub
 searchServicePrefix: https://{{GAE_VERSION}}-dot-search-dot-dartlang-pub.appspot.com
