+++ app/test/admin/moderate_package_test.dart
@@ -69,11 +69,6 @@ void main() {
 
     testWithProfile(
       'update state',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final mc = await _report('neon');
 
@@ -123,11 +118,6 @@ void main() {
 
     testWithProfile(
       'clear moderation flag',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final mc = await _report('oxygen');
         await expectModerationActions(mc.caseId, actions: []);
@@ -219,11 +209,6 @@ void main() {
 
     testWithProfile(
       'API endpoints return not found',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final jsonUrls = [
           '/api/packages/oxygen',
@@ -254,11 +239,6 @@ void main() {
 
     testWithProfile(
       'public pages are displaying moderation notice',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final htmlUrls = [
           '/packages/oxygen',
@@ -305,11 +285,6 @@ void main() {
 
     testWithProfile(
       'not included in search',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         await searchBackend.doCreateAndUpdateSnapshot(
           FakeGlobalLockClaim(clock.now().add(Duration(seconds: 3))),
@@ -354,16 +329,8 @@ void main() {
 
     testWithProfile(
       'archives are removed from public buckets',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final publicUrls = [
-          '${activeConfiguration.storageBaseUrl}'
-              '/${activeConfiguration.publicPackagesBucketName}'
-              '/packages/oxygen-1.0.0.tar.gz',
           '${activeConfiguration.storageBaseUrl}'
               '/${activeConfiguration.exportedApiBucketName}'
               '/latest/api/archives/oxygen-1.0.0.tar.gz',
@@ -389,14 +356,7 @@ void main() {
         await _moderate('oxygen', state: true, caseId: mc.caseId);
         await expectStatusCode(404);
 
-        // another check after background tasks are running
-        await packageBackend.tarballStorage.updatePublicArchiveBucket();
-        await expectStatusCode(404);
-
         await _moderate('oxygen', state: false, caseId: mc.caseId);
-        await expectStatusCode(200);
-        // another check after background tasks are running
-        await packageBackend.tarballStorage.updatePublicArchiveBucket();
         final restoredBytes = await expectStatusCode(200);
         expect(restoredBytes, bytes);
       },
@@ -405,11 +365,6 @@ void main() {
     testWithProfile(
       'analysis results are cleared',
       processJobsWithFakeRunners: true,
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.0.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         final score1 = await scoreCardBackend.getScoreCardData(
           'oxygen',
@@ -445,10 +400,6 @@ void main() {
 
     testWithProfile(
       'cleanup deletes datastore entities and canonical archive file',
-      expectedLogMessages: [
-        'SHOUT Deleting object from public bucket: "packages/oxygen-1.2.0.tar.gz".',
-        'SHOUT Deleting object from public bucket: "packages/oxygen-2.0.0-dev.tar.gz".',
-      ],
       fn: () async {
         // delete old version
         await accountBackend.withBearerToken(siteAdminToken, () async {
@@ -458,10 +409,7 @@ void main() {
 
         // canonical file is present
         expect(
-          await packageBackend.tarballStorage.getCanonicalBucketArchiveInfo(
-            'oxygen',
-            '1.2.0',
-          ),
+          await packageBackend.getCanonicalBucketArchiveInfo('oxygen', '1.2.0'),
           isNotNull,
         );
 
@@ -479,10 +427,7 @@ void main() {
           isNull,
         );
         expect(
-          await packageBackend.tarballStorage.getCanonicalBucketArchiveInfo(
-            'oxygen',
-            '1.2.0',
-          ),
+          await packageBackend.getCanonicalBucketArchiveInfo('oxygen', '1.2.0'),
           isNull,
         );
